Revert "Added event endpoint, added json middleware and refactoring (#1)"
This reverts commit 39525e500e5b31081b99dec75f41f77d97fa07b7.
