+++ backend/app.ts
@@ -1,15 +1,22 @@
+import { env } from 'process';
 import express from "express";
 import cors from 'cors';
 import compression from 'compression'
 
-import eventRouter from './routes/event';
-import corsConfig from './config/cors';
+
+let originList: any[] = ['*'] // allow all by default
+if (env.ROOT_DOMAIN !== undefined) {
+    const subdomainRe = new RegExp(`\.${env.ROOT_DOMAIN.replace(".", "\\.")}$`)
+    originList = [`http://${env.ROOT_DOMAIN}`, `https://${env.ROOT_DOMAIN}`, subdomainRe]
+}
+
+var corsOptions = { // ref: https://expressjs.com/en/resources/middleware/cors.html
+    origin: originList, 
+    optionsSuccessStatus: 200 // some legacy browsers (IE11, various SmartTVs) choke on 204
+}
 
 const app = express()
-app.use(cors(corsConfig))
-app.use(express.json({ limit: "64mb" })) 
+app.use(cors(corsOptions))
 app.use(compression())
-
-app.use('/event', eventRouter)
 
 export default app
