--- backend/controllers/event.ts
@@ -1,5 +1,0 @@
-import event from '../config/event'
-
-export const getInfo = (req: any, res: any) => {
-    return event
-}
