+++ backend/prisma/schema.prisma
@@ -11,7 +11,7 @@ datasource db {
 }
 
 model Ticket {
-  UUID String @id @default(dbgenerated("uuid_generate_v1mc()"))
+  UUID String @id @default(dbgenerated("uuid_generate_v1()"))
   ownerName String @unique
   ownerContacts String[]
   ownerAffiliation Organisation @relation(fields: [ownerAffiliationId], references: [UUID])
@@ -24,7 +24,7 @@ model Ticket {
 }
 
 model Quota {
-  UUID String @id @default(dbgenerated("uuid_generate_v1mc()"))
+  UUID String @id @default(dbgenerated("uuid_generate_v1()"))
   quotaType QuotaType @relation(fields: [quotaTypeId], references: [UUID])
   quotaTypeId String
   usageLeft Int @default(1)
@@ -33,14 +33,14 @@ model Quota {
 }
 
 model QuotaType {
-  UUID String @id @default(dbgenerated("uuid_generate_v1mc()"))
+  UUID String @id @default(dbgenerated("uuid_generate_v1()"))
   name String @unique
   description String @default("")
   quotas Quota[]
 }
 
 model Organisation {
-  UUID String @id @default(dbgenerated("uuid_generate_v1mc()"))
+  UUID String @id @default(dbgenerated("uuid_generate_v1()"))
   name String @unique
   managers User[]
   publishedInvitations Invitation[]
@@ -48,7 +48,7 @@ model Organisation {
 }
 
 model Invitation {
-  UUID String @id @default(dbgenerated("uuid_generate_v1mc()"))
+  UUID String @id @default(dbgenerated("uuid_generate_v1()"))
   publisher Organisation @relation(fields: [organisationId], references: [UUID])
   organisationId String
   usageLeft Int @default(1)
@@ -57,7 +57,7 @@ model Invitation {
 }
 
 model User {
-  UUID String @id @default(dbgenerated("uuid_generate_v1mc()"))
+  UUID String @id @default(dbgenerated("uuid_generate_v1()"))
   username String @unique
   role Int
   passwordHash String // impose max real password length=50 to be safe.
@@ -78,12 +78,12 @@ model Log {
   id Int @id @default(autoincrement())
   summary String 
   description String
-  logAction LogAction @relation(fields: [logActionId], references: [id])
-  logActionId Int
+  logType LogType @relation(fields: [logTypeId], references: [id])
+  logTypeId Int
   createdTime DateTime @default(now())
 }
 
-model LogAction {
+model LogType {
   id Int @id @default(autoincrement())
   name String @unique
   logs Log[]
