--- backend/routes/event.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { getInfo } from "../controllers/event";
-
-const eventRouter = Router({})
-
-eventRouter.get('/', getInfo)
-
-export default eventRouter
